+++ CHANGELOG.md
@@ -116,6 +116,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   - Removed weak reference back to Instrument within Files class
   - Fixed access of xarray data with more than one dimension (#471)
   - Improved robustness of eval(inst.__repr__()) (#636)
+  - Fixed `calc_solar_local_time` for data sets with longitude coordinates
 - Maintenance
   - nose dependency removed from unit tests
   - Specified `dtype` for empty pandas.Series for forward compatibility